cleanup firstrun
