+++ .d.ts
@@ -2,3 +2,8 @@ declare module '*.png' {
   const value: any
   export default value
 }
+
+declare module '*.md' {
+  const value: any
+  export default value
+}
