+++ legacy/helpers/appSettings.js
@@ -2,42 +2,8 @@
 
 import ExtendableError from 'es6-error'
 import _ from 'lodash'
-import loadLocalResource from 'react-native-local-resource'
 
-import CS_CZ_TOS from '../i18n/locales/terms-of-use/ada/cs-CZ.md'
-import DE_DE_TOS from '../i18n/locales/terms-of-use/ada/de-DE.md'
-import EN_US_TOS from '../i18n/locales/terms-of-use/ada/en-US.md'
-import ES_ES_TOS from '../i18n/locales/terms-of-use/ada/es-ES.md'
-import FR_FR_TOS from '../i18n/locales/terms-of-use/ada/fr-FR.md'
-import HU_HU_TOS from '../i18n/locales/terms-of-use/ada/hu-HU.md'
-import ID_ID_TOS from '../i18n/locales/terms-of-use/ada/id-ID.md'
-import IT_IT_TOS from '../i18n/locales/terms-of-use/ada/it-IT.md'
-import JA_JP_TOS from '../i18n/locales/terms-of-use/ada/ja-JP.md'
-import KO_KR_TOS from '../i18n/locales/terms-of-use/ada/ko-KR.md'
-import NL_NL_TOS from '../i18n/locales/terms-of-use/ada/nl-NL.md'
-import PT_BR_TOS from '../i18n/locales/terms-of-use/ada/pt-BR.md'
-import RU_RU_TOS from '../i18n/locales/terms-of-use/ada/ru-RU.md'
-import SK_SK_TOS from '../i18n/locales/terms-of-use/ada/sk-SK.md'
-import ZH_CN_TOS from '../i18n/locales/terms-of-use/ada/zh-CN.md'
 import storage from '../utils/storage'
-
-const tosByCode = {
-  'en-US': EN_US_TOS,
-  'ja-JP': JA_JP_TOS,
-  'ko-KR': KO_KR_TOS,
-  'ru-RU': RU_RU_TOS,
-  'es-ES': ES_ES_TOS,
-  'zh-Hans': ZH_CN_TOS,
-  'id-ID': ID_ID_TOS,
-  'pt-BR': PT_BR_TOS,
-  'de-DE': DE_DE_TOS,
-  'fr-FR': FR_FR_TOS,
-  'it-IT': IT_IT_TOS,
-  'nl-NL': NL_NL_TOS,
-  'cs-CZ': CS_CZ_TOS,
-  'hu-HU': HU_HU_TOS,
-  'sk-SK': SK_SK_TOS,
-}
 
 // Note(ppershing): following values have to be in sync with
 // keys in redux state
@@ -81,10 +47,4 @@ export const readAppSettings = async () => {
     const setting = _.last(key.split('/'))
     return {...acc, [setting]: value}
   }, {})
-}
-
-export const loadTOS = async (languageCode: string) => {
-  const tosFile = tosByCode[languageCode]
-  const tos = await loadLocalResource(tosFile)
-  return tos
 }
