--- src/FirstRun/AcceptTermsOfService.stories.tsx
@@ -1,6 +1,0 @@
-import {storiesOf} from '@storybook/react-native'
-import React from 'react'
-
-import {AcceptTermsOfServiceScreen} from './AcceptTermsOfServiceScreen'
-
-storiesOf('AcceptTermsOfServiceScreen', module).add('Default', () => <AcceptTermsOfServiceScreen />)
