+++ src/FirstRun/CustomPinScreen/CustomPinScreen.tsx
@@ -4,10 +4,10 @@ import {StyleSheet} from 'react-native'
 import {SafeAreaView} from 'react-native-safe-area-context'
 import {useDispatch, useSelector} from 'react-redux'
 
-import {encryptAndStoreCustomPin, setSystemAuth, signin} from '../../legacy/actions'
-import PinRegistrationForm from '../../legacy/components/Common/PinRegistrationForm'
-import {StatusBar} from '../../legacy/components/UiKit'
-import {isAuthenticatedSelector} from '../../legacy/selectors'
+import {encryptAndStoreCustomPin, setSystemAuth, signin} from '../../../legacy/actions'
+import PinRegistrationForm from '../../../legacy/components/Common/PinRegistrationForm'
+import {StatusBar} from '../../../legacy/components/UiKit'
+import {isAuthenticatedSelector} from '../../../legacy/selectors'
 
 export const CustomPinScreen = () => {
   const strings = useStrings()
