+++ src/FirstRun/LanguagePickerScreen/LanguagePickerScreen.stories.tsx
@@ -1,8 +1,8 @@
 import {storiesOf} from '@storybook/react-native'
 import React from 'react'
 
-import {ReduxProvider} from '../../storybook/ReduxProvider'
-import {LanguagePickerScreen} from './LanguagePickerScreen'
+import {ReduxProvider} from '../../../storybook/ReduxProvider'
+import {LanguagePickerScreen} from '.'
 
 storiesOf('LanguagePickerScreen', module).add('Default', () => (
   <ReduxProvider>
