+++ src/FirstRun/LanguagePickerScreen/LanguagePickerScreen.tsx
@@ -4,10 +4,10 @@ import {StyleSheet} from 'react-native'
 import {SafeAreaView} from 'react-native-safe-area-context'
 import {useDispatch, useSelector} from 'react-redux'
 
-import {changeAndSaveLanguage, changeLanguage} from '../../legacy/actions/language'
-import LanguagePicker from '../../legacy/components/Common/LanguagePicker'
-import {FIRST_RUN_ROUTES} from '../../legacy/RoutesList'
-import {languageSelector} from '../../legacy/selectors'
+import {changeAndSaveLanguage, changeLanguage} from '../../../legacy/actions/language'
+import LanguagePicker from '../../../legacy/components/Common/LanguagePicker'
+import {FIRST_RUN_ROUTES} from '../../../legacy/RoutesList'
+import {languageSelector} from '../../../legacy/selectors'
 
 export const LanguagePickerScreen = () => {
   const navigation = useNavigation()
