+++ src/FirstRun/TermsOfServiceScreen/TermsOfService/TermsOfService.tsx
@@ -1,14 +1,12 @@
-// @flow
-
 import React from 'react'
 import {ActivityIndicator} from 'react-native'
 import Markdown from 'react-native-easy-markdown'
 import {useSelector} from 'react-redux'
 
-import {loadTOS} from '../../helpers/appSettings'
-import {languageSelector} from '../../selectors'
+import {languageSelector} from '../../../../legacy/selectors'
+import {loadTOS} from './loadTos'
 
-const TermsOfService = () => {
+export const TermsOfService = () => {
   const [tos, setTos] = React.useState()
   const languageCode = useSelector(languageSelector)
 
@@ -18,5 +16,3 @@ const TermsOfService = () => {
 
   return tos ? <Markdown>{tos}</Markdown> : <ActivityIndicator size={'large'} color={'black'} />
 }
-
-export default TermsOfService
