+++ src/FirstRun/TermsOfServiceScreen/TermsOfService/index.tsx
@@ -1,0 +1,1 @@
+export * from './TermsOfService'
