+++ src/FirstRun/TermsOfServiceScreen/TermsOfService/tos/cs-CZ.md
