+++ src/FirstRun/TermsOfServiceScreen/TermsOfService/tos/fr-FR.md
