+++ src/FirstRun/TermsOfServiceScreen/TermsOfService/tos/hu-HU.md
