+++ src/FirstRun/TermsOfServiceScreen/TermsOfService/tos/ja-JP.md
