+++ src/FirstRun/TermsOfServiceScreen/TermsOfService/tos/ru-RU.md
