+++ src/FirstRun/TermsOfServiceScreen/TermsOfService/tos/sk-SK.md
