+++ src/FirstRun/TermsOfServiceScreen/TermsOfServiceScreen.stories.tsx
@@ -1,0 +1,6 @@
+import {storiesOf} from '@storybook/react-native'
+import React from 'react'
+
+import {TermsOfServiceScreen} from './TermsOfServiceScreen'
+
+storiesOf('TermsOfServiceScreen', module).add('Default', () => <TermsOfServiceScreen />)
